requirements.txt
pymilvus>=2.3.0
ray[default]>=2.5.0
pypdf>=3.17.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
langchain>=0.0.350
transformers>=4.30.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
